+++ src/gromo/modules/conv2d_growing_module.py
@@ -501,6 +501,8 @@ class Conv2dGrowingModule(GrowingModule):
         device for the layer
     name: str | None
         name of the layer used for debugging purpose
+    target_in_channels: int | None
+        target number of input channels for the layer when growing is performed
     """
 
     def __init__(
@@ -521,6 +523,7 @@ class Conv2dGrowingModule(GrowingModule):
         allow_growing: bool = False,
         device: torch.device | None = None,
         name: str | None = None,
+        target_in_channels: int | None = None,
     ) -> None:
         if isinstance(kernel_size, int):
             kernel_size = (kernel_size, kernel_size)
@@ -550,6 +553,8 @@ class Conv2dGrowingModule(GrowingModule):
             ),
             device=device,
             name=name,
+            target_in_neurons=target_in_channels,
+            initial_in_neurons=in_channels,
         )
         self.layer: torch.nn.Conv2d
         self.kernel_size = self.layer.kernel_size
@@ -564,6 +569,10 @@ class Conv2dGrowingModule(GrowingModule):
     # this function is used to estimate the F.O. improvement of the loss after the
     # extension of the network however this won't work if we do not have only the
     # activation function as the post_layer_function
+
+    @property
+    def in_neurons(self) -> int:
+        return self.in_channels
 
     @property
     def in_channels(self):
@@ -1139,6 +1148,7 @@ class RestrictedConv2dGrowingModule(Conv2dGrowingModule):
         allow_growing: bool = False,
         device: torch.device | None = None,
         name: str | None = None,
+        target_in_channels: int | None = None,
     ) -> None:
         super(RestrictedConv2dGrowingModule, self).__init__(
             in_channels=in_channels,
@@ -1156,6 +1166,7 @@ class RestrictedConv2dGrowingModule(Conv2dGrowingModule):
             allow_growing=allow_growing,
             device=device,
             name=name,
+            target_in_channels=target_in_channels,
         )
         self.bordering_convolution = None
 
@@ -1553,6 +1564,7 @@ class FullConv2dGrowingModule(Conv2dGrowingModule):
         allow_growing: bool = False,
         device: torch.device | None = None,
         name: str | None = None,
+        target_in_channels: int | None = None,
     ) -> None:
         super(FullConv2dGrowingModule, self).__init__(
             in_channels=in_channels,
@@ -1570,6 +1582,7 @@ class FullConv2dGrowingModule(Conv2dGrowingModule):
             allow_growing=allow_growing,
             device=device,
             name=name,
+            target_in_channels=target_in_channels,
         )
         self._mask_tensor_t: torch.Tensor | None = None
         self._tensor_s_growth = TensorStatistic(