+++ src/gromo/modules/growing_module.py
@@ -552,6 +552,8 @@ class GrowingModule(torch.nn.Module):
         next_module: torch.nn.Module | None = None,
         device: torch.device | None = None,
         name: str | None = None,
+        target_in_neurons: int | None = None,
+        initial_in_neurons: int | None = None,
     ) -> None:
         """
         Initialize a GrowingModule.
@@ -576,6 +578,11 @@ class GrowingModule(torch.nn.Module):
             device to use
         name: str | None
             name of the module
+        target_in_neurons: int | None
+            target number of input neurons for the layer at the end of the growth process
+        initial_in_neurons: int | None
+            initial number of input neurons for the layer at the beginning of the growth
+            process
         """
         if tensor_s_shape is None:
             warnings.warn(
@@ -599,6 +606,8 @@ class GrowingModule(torch.nn.Module):
             if name is None
             else f"{self.__class__.__name__}({name})"
         )
+        self.target_in_neurons = target_in_neurons
+        self._initial_in_neurons = initial_in_neurons
         self._config_data, _ = load_config()
         self.device = get_correct_device(self, device)
 
@@ -697,6 +706,10 @@ class GrowingModule(torch.nn.Module):
             device=self.device,
             name=f"C({self.name})",
         )
+
+    @property
+    def in_neurons(self) -> int:
+        raise NotImplementedError
 
     @property
     def in_features(self) -> int:
@@ -870,9 +883,9 @@ class GrowingModule(torch.nn.Module):
         if key == "store_input" and value is not self.store_input:
             self.__dict__["store_input"] = value
             if isinstance(self.previous_module, MergeGrowingModule):
-                # As a MergeGrowingModule may have multiple next modules
-                # we need to keep track of the number of modules that require the activity
-                # to be stored. Hence we store it as long as one of the module requires it.
+                # As a MergeGrowingModule may have multiple next modules we need to
+                # keep track of the number of modules that require the activity to be
+                # stored. Hence we store it as long as one of the module requires it.
                 self.previous_module.store_activity += 1 if value else -1
             else:
                 self._internal_store_input = value
@@ -905,7 +918,8 @@ class GrowingModule(torch.nn.Module):
                 pass
             else:
                 raise TypeError(
-                    f"Previous module must be a GrowingModule or MergeGrowingModule, got {type(self.previous_module)}"
+                    f"Previous module must be a GrowingModule or MergeGrowingModule, "
+                    f"got {type(self.previous_module)}"
                 )
         elif key == "weight":
             self.layer.weight = value
@@ -1006,7 +1020,8 @@ class GrowingModule(torch.nn.Module):
             else:
                 if x_ext is not None:  # TODO: and is not empty
                     warnings.warn(
-                        f"x_ext must be None got {x_ext} for {self.name}. As the input is not extended, no extension is needed.",
+                        f"x_ext must be None got {x_ext} for {self.name}. As the input "
+                        f"is not extended, no extension is needed.",
                         UserWarning,
                     )
 
@@ -1031,7 +1046,8 @@ class GrowingModule(torch.nn.Module):
         force_update: bool = True,
     ) -> tuple[int, ...] | None:
         """
-        Update the input size of the layer. Either according to the parameter or the input currently stored.
+        Update the input size of the layer. Either according to the parameter or
+        the input currently stored.
 
         Parameters
         ----------
@@ -1120,7 +1136,8 @@ class GrowingModule(torch.nn.Module):
     # Statistics computation
     def projected_v_goal(self, input_vector: torch.Tensor) -> torch.Tensor:
         """
-        Compute the projected gradient of the goal with respect to the activity of the layer.
+        Compute the projected gradient of the goal with respect to the activity
+        of the layer.
 
         dLoss/dA_proj := dLoss/dA - dW B[-1] where A is the pre-activation vector of the
         layer, and dW the optimal delta for the layer
@@ -1184,12 +1201,13 @@ class GrowingModule(torch.nn.Module):
             return self.previous_module.tensor_s
         elif isinstance(self.previous_module, MergeGrowingModule):
             raise NotImplementedError(
-                f"S growth is not implemented for module preceded by an MergeGrowingModule."
-                f" (error in {self.name})"
+                f"S growth is not implemented for module preceded by an "
+                f"MergeGrowingModule. (error in {self.name})"
             )
         else:
             raise NotImplementedError(
-                f"S growth is not implemented yet for {type(self.previous_module)} as previous module."
+                f"S growth is not implemented yet for {type(self.previous_module)} "
+                f"as previous module."
             )
 
     @tensor_s_growth.setter
@@ -1199,7 +1217,8 @@ class GrowingModule(torch.nn.Module):
         """
         raise AttributeError(
             f"You tried to set tensor_s_growth of a GrowingModule (name={self.name})."
-            "This is not allowed because tensor_s_growth refers to the previous module's tensor_s, not the current module's tensor_s."
+            f"This is not allowed because tensor_s_growth refers to the previous module's"
+            f" tensor_s, not the current module's tensor_s."
         )
 
     def compute_m_update(
@@ -1506,8 +1525,8 @@ class GrowingModule(torch.nn.Module):
                 < 1e-4
             ):
                 warnings.warn(
-                    f"Scaling factor {scaling_factor} is different from the one "
-                    f"used during the extended_forward {self._scaling_factor_next_module}."
+                    f"Scaling factor {scaling_factor} is different from the one used"
+                    f" during the extended_forward {self._scaling_factor_next_module}."
                 )
         if extension_size > 0 or self.extended_output_layer is not None:
             assert isinstance(self.extended_output_layer, torch.nn.Module), (
@@ -1588,7 +1607,8 @@ class GrowingModule(torch.nn.Module):
                 ), "The bias of the input extension must be null."
                 if self.scaling_factor == 0:
                     warnings.warn(
-                        "The scaling factor is null. The input extension will have no effect."
+                        "The scaling factor is null. "
+                        "The input extension will have no effect."
                     )
                 self.layer_in_extension(
                     weight=sqrt_factor * self.extended_input_layer.weight
@@ -1600,8 +1620,8 @@ class GrowingModule(torch.nn.Module):
                         if extension_size is None:
                             assert self.eigenvalues_extension is not None, (
                                 "We need to determine the size of the extension but "
-                                "it was not given as parameter nor could be automatically "
-                                "determined as self.eigenvalues_extension is None"
+                                "it was not given as parameter nor could be automatically"
+                                " determined as self.eigenvalues_extension is None"
                                 f"(Error occurred in {self.name})"
                             )
                             extension_size = self.eigenvalues_extension.shape[0]
@@ -1657,7 +1677,8 @@ class GrowingModule(torch.nn.Module):
         Returns
         -------
         tuple[torch.Tensor, torch.Tensor | None, torch.Tensor | float]
-            optimal delta for the weights, the biases if needed and the first order decrease
+            optimal delta for the weights, the biases if needed and
+            the first order decrease
         """
         tensor_s = self.tensor_s()
         tensor_m = self.tensor_m()
@@ -1693,7 +1714,8 @@ class GrowingModule(torch.nn.Module):
         Parameters
         ----------
         numerical_threshold: float
-            threshold to consider an eigenvalue as zero in the square root of the inverse of S
+            threshold to consider an eigenvalue as zero in the square root of
+            the inverse of S
         statistical_threshold: float
             threshold to consider an eigenvalue as zero in the SVD of S{-1/2} N
         maximum_added_neurons: int | None
@@ -1756,7 +1778,8 @@ class GrowingModule(torch.nn.Module):
         Parameters
         ----------
         numerical_threshold: float
-            threshold to consider an eigenvalue as zero in the square root of the inverse of S
+            threshold to consider an eigenvalue as zero in the square root of
+            the inverse of S
         statistical_threshold: float
             threshold to consider an eigenvalue as zero in the SVD of S{-1/2} N
         maximum_added_neurons: int | None
@@ -1812,7 +1835,8 @@ class GrowingModule(torch.nn.Module):
         Parameters
         ----------
         numerical_threshold: float
-            threshold to consider an eigenvalue as zero in the square root of the inverse of S
+            threshold to consider an eigenvalue as zero in the square root of
+            the inverse of S
         statistical_threshold: float
             threshold to consider an eigenvalue as zero in the SVD of S{-1/2} N
         maximum_added_neurons: int | None
@@ -1934,9 +1958,11 @@ class GrowingModule(torch.nn.Module):
         Raises
         ------
         NotImplementedError
-            raised when include_previous is True and the previous module is of type MergeGrowingModule
+            raised when include_previous is True and the previous module is
+            of type MergeGrowingModule
         TypeError
-            raised when the previous module is not of type GrowingModule or MergeGrowingModule
+            raised when the previous module is not of type GrowingModule
+            or MergeGrowingModule
         """
         if delete_delta:
             self.optimal_delta_layer = None
@@ -1988,7 +2014,8 @@ class GrowingModule(torch.nn.Module):
                                 "This may lead to errors when using extended_forward.",
                                 UserWarning,
                             )
-                        # otherwise it is ok as user already deleted the extended_output_layer
+                        # otherwise it is ok as user already deleted
+                        # the extended_output_layer
                     elif isinstance(self.previous_module, MergeGrowingModule):
                         return
                         # the user intentionally decided to take care of deletion of the
@@ -2457,6 +2484,81 @@ class GrowingModule(torch.nn.Module):
                 self.previous_module.bias,
                 self.previous_module.get_fan_in_from_layer(layer_to_init),
             )
+
+    def missing_neurons(self) -> int:
+        """
+        Get the number of missing neurons to reach the target hidden features.
+
+        Returns
+        -------
+        int
+            number of missing neurons
+        """
+        if self.target_in_neurons is None:
+            raise ValueError(
+                "Target in neurons is not set, cannot compute missing neurons."
+            )
+        return self.target_in_neurons - self.in_neurons
+
+    def number_of_neurons_to_add(
+        self,
+        method: str = "fixed_proportional",
+        number_of_growth_steps: int = 1,
+    ) -> int:
+        """Get the number of neurons to add in the next growth step.
+
+        Methods
+        -------
+        - fixed_proportional: add a fixed proportion of the total number of neurons
+          to add at each growth step. The amount to add is computed as
+          an integer division as a consequence a few neurons may remain to be added
+          after all growth steps have been performed.
+
+
+        Parameters
+        ----------
+        method : str
+            Method to use for determining the number of neurons to add.
+            Options are "fixed_proportional".
+        number_of_growth_steps : int
+            Number of growth steps planned, used only if method is "fixed_proportional".
+
+        Returns
+        -------
+        int
+            Number of neurons to add.
+        """
+        if method == "fixed_proportional":
+            if self.target_in_neurons is None:
+                raise ValueError(
+                    "Target in neurons is not set, cannot compute neurons to add."
+                )
+            if self._initial_in_neurons is None:
+                raise ValueError(
+                    "Initial in neurons is not set, cannot compute neurons to add."
+                )
+            total_to_add = self.target_in_neurons - self._initial_in_neurons
+            return total_to_add // number_of_growth_steps
+        else:
+            raise ValueError(f"Unknown method: {method}.")
+
+    def complete_growth(self, extension_kwargs: dict) -> None:
+        """
+        Complete the growth to the target size.
+
+        Parameters
+        ----------
+        extension_kwargs : dict
+            Additional arguments for creating layer extensions.
+        """
+        neurons_to_add = self.missing_neurons()
+        if neurons_to_add > 0:
+            self.create_layer_extensions(
+                extension_size=neurons_to_add,
+                **extension_kwargs,
+            )
+            self.apply_change(extension_size=neurons_to_add)
+            self.delete_update(include_previous=True)
 
 
 if __name__ == "__main__":